decoupled handlers, events. added utils
